Move _prepare_dataframe to sc.get
