+++ scanpy/get.py
@@ -1,11 +1,15 @@
 """This module contains helper functions for accessing data."""
-from typing import Optional, Iterable, Tuple
+from typing import Optional, Iterable, Tuple, Mapping, Union, Sequence
 
 import numpy as np
 import pandas as pd
-from scipy.sparse import spmatrix, issparse
+from pandas.api.types import is_categorical_dtype
+from scipy.sparse import spmatrix
 
 from anndata import AnnData
+from ._utils import sanitize_anndata
+
+_VarNames = Union[str, Sequence[str]]
 
 # --------------------------------------------------------------------------------
 # Plotting data helpers
@@ -346,3 +350,149 @@ def _set_obs_rep(adata, val, *, use_raw=False, layer=None, obsm=None, obsp=None)
             "That was unexpected. Please report this bug at:\n\n\t"
             " https://github.com/theislab/scanpy/issues"
         )
+
+
+def _prepare_dataframe(
+    adata: AnnData,
+    var_names: Optional[Union[_VarNames, Mapping[str, _VarNames]]] = None,
+    groupby: Optional[Union[str, Sequence[str]]] = None,
+    use_raw: Optional[bool] = None,
+    log: bool = False,
+    num_categories: int = 7,
+    layer: Optional[str] = None,
+    gene_symbols: Optional[str] = None,
+    concat_indices: bool = True,
+):
+    """
+    Given the anndata object, prepares a data frame in which the row index are the categories
+    defined by group by and the columns correspond to var_names.
+
+    Parameters
+    ----------
+    adata
+        Annotated data matrix.
+    var_names
+        `var_names` should be a valid subset of `adata.var_names`. All genes are used if no
+        given.
+    groupby
+        The key of the observation grouping to consider. It is expected that
+        groupby is a categorical. If groupby is not a categorical observation,
+        it would be subdivided into `num_categories`.
+    use_raw
+        Use `raw` attribute of `adata` if present.
+    log
+        Use the log of the values
+    num_categories
+        Only used if groupby observation is not categorical. This value
+        determines the number of groups into which the groupby observation
+        should be subdivided.
+    gene_symbols
+        Key for field in .var that stores gene symbols.
+    concat_indices
+        Concatenates categorical indices into a single categorical index, if 
+        groupby is a sequence. True by default.
+
+    Returns
+    -------
+    Tuple of `pandas.DataFrame` and list of categories.
+    """
+    from scipy.sparse import issparse
+
+    sanitize_anndata(adata)
+    if use_raw is None and adata.raw is not None:
+        use_raw = True
+    if isinstance(var_names, str):
+        var_names = [var_names]
+    if var_names is None:
+        if use_raw:
+            var_names = adata.raw.var_names.values
+        else:
+            var_names = adata.var_names.values
+
+    if groupby is not None:
+        if isinstance(groupby, str):
+            # if not a list, turn into a list
+            groupby = [groupby]
+        for group in groupby:
+            if group not in adata.obs_keys():
+                raise ValueError(
+                    'groupby has to be a valid observation. '
+                    f'Given {group}, is not in observations: {adata.obs_keys()}'
+                )
+
+    if gene_symbols is not None and gene_symbols in adata.var.columns:
+        # translate gene_symbols to var_names
+        # slow method but gives a meaningful error if no gene symbol is found:
+        translated_var_names = []
+        # if we're using raw to plot, we should also do gene symbol translations
+        # using raw
+        if use_raw:
+            adata_or_raw = adata.raw
+        else:
+            adata_or_raw = adata
+        for symbol in var_names:
+            if symbol not in adata_or_raw.var[gene_symbols].values:
+                logg.error(
+                    f"Gene symbol {symbol!r} not found in given "
+                    f"gene_symbols column: {gene_symbols!r}"
+                )
+                return
+            translated_var_names.append(
+                adata_or_raw.var[adata_or_raw.var[gene_symbols] == symbol].index[0]
+            )
+        symbols = var_names
+        var_names = translated_var_names
+    if layer is not None:
+        if layer not in adata.layers.keys():
+            raise KeyError(
+                f'Selected layer: {layer} is not in the layers list. '
+                f'The list of valid layers is: {adata.layers.keys()}'
+            )
+        matrix = adata[:, var_names].layers[layer]
+    elif use_raw:
+        matrix = adata.raw[:, var_names].X
+    else:
+        matrix = adata[:, var_names].X
+
+    if issparse(matrix):
+        matrix = matrix.toarray()
+    if log:
+        matrix = np.log1p(matrix)
+
+    obs_tidy = pd.DataFrame(matrix, columns=var_names)
+    if groupby is None:
+        groupby = ''
+        obs_tidy_idx = pd.Series(np.repeat('', len(obs_tidy))).astype('category')
+    else:
+        if len(groupby) == 1 and not is_categorical_dtype(adata.obs[groupby[0]]):
+            # if the groupby column is not categorical, turn it into one
+            # by subdividing into  `num_categories` categories
+            obs_tidy_idx = pd.cut(adata.obs[groupby[0]], num_categories)
+            idx_categories = obs_tidy_idx.cat.categories
+        else:
+            assert all(is_categorical_dtype(adata.obs[group]) for group in groupby)
+            if concat_indices:
+                obs_tidy_idx = adata.obs[groupby[0]]
+                if len(groupby) > 1:
+                    for group in groupby[1:]:
+                        # create new category by merging the given groupby categories
+                        obs_tidy_idx = (
+                            obs_tidy_idx.astype(str) + "_" + adata.obs[group].astype(str)
+                        ).astype('category')
+                obs_tidy_idx.name = "_".join(groupby)
+                idx_categories = obs_tidy_idx.cat.categories
+            else:
+                obs_tidy_idx = [adata.obs[group] for group in groupby] # keep as multiindex
+                idx_categories = [x.cat.categories for x in obs_tidy_idx]
+
+    obs_tidy.set_index(obs_tidy_idx, inplace=True)
+    if gene_symbols is not None:
+        # translate the column names to the symbol names
+        obs_tidy.rename(
+            columns={var_names[x]: symbols[x] for x in range(len(var_names))},
+            inplace=True,
+        )
+
+    return idx_categories, obs_tidy
+
+
